% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments.R
\docType{data}
\name{assigned_data}
\alias{assigned_data}
\title{assignments on 8 data files}
\format{a list with 8 entries, each of which has:
\describe{
\item{tic}{the total intensity of peaks in the spectrum}
\item{assignments}{a data.frame of assignments}
\item{sample}{which sample the assignment are from}
}}
\source{
generated by \code{read_smirfe_assignments} on 8 sample files.
}
\usage{
assigned_data
}
\description{
A dataset containing SMIRFE assignments for 8 samples.
}
\keyword{datasets}
