% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{calculate_assignment_scores}
\alias{calculate_assignment_scores}
\title{calculate assignment scores}
\usage{
calculate_assignment_scores(single_sample, variable = "e_value",
  weight = 1, calculation = "one_minus")
}
\arguments{
\item{variable}{the variable to calculate scores from}

\item{weight}{multiplier to apply to the scores (default = 1)}

\item{calculation}{how to calculate the scores (default = "one_minus")}

\item{sample_assignments}{the assignments data.frame}
}
\value{
data.frame of assignments with scores added
}
\description{
Given the assignment data.frame, calculate new scores for every assignment.
The default is 1 - e_value, with a weighting of 1.
}
