% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_changes.R
\name{calculate_mean_fc}
\alias{calculate_mean_fc}
\title{calculate mean fold changes}
\usage{
calculate_mean_fc(intensities, grouped_samples, add_min = 1e-12)
}
\arguments{
\item{intensities}{matrix of intensities, rows are features, columns are samples}

\item{grouped_samples}{list of samples}

\item{add_min}{a minimum value to add to 0 so that log doesn't complain}
}
\value{
data.frame
}
\description{
Provided values of intensities and groups of samples, calculates log2 fold-changes
between the mean values across the groups of samples.
}
