% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_counts.R
\name{check_imf_emf_elemental_counts}
\alias{check_imf_emf_elemental_counts}
\title{Compare IMF and EMF Counts}
\usage{
check_imf_emf_elemental_counts(imf_counts, emf_counts)
}
\arguments{
\item{imf_counts}{data.frame from \link{get_imf_elements}}

\item{emf_counts}{data.frame from \link{get_emf_elements}}
}
\value{
logical
}
\description{
Given an IMF counts \code{data.frame} and an EMF counts \code{data.frame}, compare them
to make sure that they contain identical counts of each element.
}
