% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{choose_emf}
\alias{choose_emf}
\title{choose most likely elemental molecular formula's}
\usage{
choose_emf(grouped_emfs, peak_mz, keep_ratio = 0.9)
}
\arguments{
\item{grouped_emfs}{list of grouped EMFs}

\item{peak_mz}{data.frame of peak M/Zs}

\item{keep_ratio}{how close to the maximum voted EMF to keep other things? (default is 0.9)}
}
\value{
data.frame
}
\description{
Given a list of grouped EMFs (assumed to be from a single sudo EMF), votes on the most likely
EMF based on the sum of \code{1 - e_value} across the grouped EMFs. If a grouped EMF does not
have a matching EMF, then attempts to match peaks to the voted EMFs based on M/Z and matching
to EMF peaks in natural abundance order.
}
