% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_assignments.R
\name{choose_imfs}
\alias{choose_imfs}
\title{decide on IMFs}
\usage{
choose_imfs(assignment_data, sample_peak = "sample_peak", imf = "IMF",
  sample = "sample", e_value = "IMF.E.Value", min_e_value = 1e-18)
}
\arguments{
\item{assignment_data}{data.frame of assignment data}

\item{sample_peak}{what defines a sample_peak combination}

\item{imf}{which is the imf variable}

\item{sample}{what is the sample}

\item{e_value}{which is the e-value column?}

\item{min_e_value}{what is the minimum e-value to use?}
}
\description{
given a set of assignment data, decide which set of IMFs should
be reported.
}
