% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_assignments.R
\name{choose_peak}
\alias{choose_peak}
\title{choose a single peak}
\usage{
choose_peak(peak_frame)
}
\arguments{
\item{peak_frame}{the data.frame with the IMF information}
}
\value{
data.frame with only a single peak
}
\description{
Need to choose a single peak to associate with each IMF. This tries to make
a choice either by assignment type, IMF E value, or M/Z difference, in that
order.
}
