% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_assignments.R
\name{choose_single_peak}
\alias{choose_single_peak}
\title{choose a peak}
\usage{
choose_single_peak(assignment_data, keep_imf, imf_mz = NULL, imf = "IMF",
  sample_peak = "sample_peak", sample = "sample")
}
\arguments{
\item{assignment_data}{the data.frame of assignment data}

\item{keep_imf}{which IMF was chosen from the assignments}

\item{imf_mz}{the assigned IMF M/Z}

\item{imf}{which variable are the IMFs listed}

\item{sample_peak}{which variable contains the sample peak information}

\item{sample}{which variable has the sample information}
}
\value{
data.frame of assignment data, with a single peak in each sample.
}
\description{
When there are multiple peaks in a sample, we need to choose one based on
the chosen representative IMF. See \strong{Details} for more information on how
that decision is made.
}
\details{
Given the \code{assignment_data} and which IMF to keep, picks which
peak should be kept for a given sample. This is made in order of:
\enumerate{
\item Does one of the peaks have a \strong{Primary} assignment matching the IMF?
\item Does one of the peaks have a \strong{Secondary} assignment matching the IMF?
\item Which peak has the \strong{smallest} M/Z difference to the chosen IMF?
}
}
