% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfs.R
\name{cleanup_na_categories}
\alias{cleanup_na_categories}
\title{Cleanup NA Categories}
\usage{
cleanup_na_categories(grouped_emf)
}
\arguments{
\item{grouped_emf}{a data.frame of peaks for the grouped emf}
}
\value{
data.frame
}
\description{
After choosing a set of peaks for a grouped EMF, it may be that there are
predicted categories that have an NA match in the Category based on formula
matching. In this particular case, we actually want to keep just the ones
that are NOT NA in both. This function does that. It also checks that we
have the same number of peaks as went in. Otherwise, it just returns the
original data.
}
