% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_assignments.R
\name{complete_emf_mappings}
\alias{complete_emf_mappings}
\title{Complete EMF mapping}
\usage{
complete_emf_mappings(assigned_data, remove_S = TRUE, progress = TRUE)
}
\arguments{
\item{assigned_data}{lists of assigned data}

\item{remove_S}{should formulas containing "S" be removed}

\item{progress}{show a progress bar for the main worker}
}
\value{
data.frame
}
\description{
Everything is reported by \code{complete_EMF} or \code{complete_IMF}, which includes
the adducts, to make sure we have unique peaks. However, for some other
work, we actually want the \strong{non-complete} EMF and IMF. This function
generates the mapping of \code{complete_EMF} to \code{isotopologue_EMF / IMF} for
other uses.
}
