% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confident_sd.R
\name{create_mz_diffs}
\alias{create_mz_diffs}
\title{create M/Z differences}
\usage{
create_mz_diffs(coefficient_info, frequency_offset = NULL,
  mz_values = seq(150, 1600, 0.5))
}
\arguments{
\item{coefficient_info}{list of coefficients from \code{get_sample_coefficients}}

\item{frequency_offset}{the offset to use, based on \code{find_confident_frequency_sd}}

\item{mz_values}{a vector of M/Z values to create offsets for}
}
\value{
data.frame with Index and Value
}
\description{
Given a list of coefficient data from multiple samples and a previously
derived \code{frequency_offset}, this generates a data.frame of M/Z locations
and differences.
}
