% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_samples.R
\name{create_sample_tracking}
\alias{create_sample_tracking}
\title{create sample tracking}
\usage{
create_sample_tracking(sample_list)
}
\arguments{
\item{sample_list}{list of samples}

\item{contains_paths}{do the samples contain full paths}
}
\value{
data.frame
}
\description{
Create a data.frame for sample tracking and filtering.
}
