% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{create_sudo_emfs}
\alias{create_sudo_emfs}
\title{create sudo elemental molecular formulas}
\usage{
create_sudo_emfs(gemf_2_emf)
}
\arguments{
\item{gemf_2_emf}{list of grouped EMFs to EMFs}
}
\value{
list
}
\description{
Given the grouped EMFs in a sample to EMFs, come up with \code{sudo} EMFs, that is groups
of shared EMFs within and across samples, based on shared EMFs.
}
\details{
This is achieved by repeated set intersection / set union cycles, where
they are repeated until the set of things is consistent between cycles, and there
are no entries to pull from.
}
