% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfs.R
\name{decide_emf_peaks}
\alias{decide_emf_peaks}
\title{Decide EMF Peaks}
\usage{
decide_emf_peaks(emf_peaks)
}
\arguments{
\item{emf_peaks}{\emph{all} peaks for a grouped EMF, as a data.frame. See Details.}
}
\value{
data.frame
}
\description{
A given grouped EMF might have several possible peaks and complete EMF possibilities.
This function attempts to narrow those down by considering if there is
a lipid category prediction, and then the difference between the maximum
number of C13 incorporations observed and the actual number of peaks
in the data for that EMF. Only those EMFs that have the same number of
peaks are returned.
}
