% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_assignments.R
\name{extract_assigned_data}
\alias{extract_assigned_data}
\title{extract data}
\usage{
extract_assigned_data(assigned_data, remove_seconary = TRUE,
  sample_peak = "sample_peak", imf = "IMF", e_value = "E.Value",
  emf = "EMF", sample = "sample", observed_mz = "ObservedMZ.Mean",
  assigned_mz = "Assigned.M.Z", height = "Height.Mean",
  other_cols = c("Adduct_IMF", "Adduct", "EMF2"), progress = TRUE)
}
\arguments{
\item{assigned_data}{a list of assignments created from \code{read_smirfe_assignments}}

\item{sample_peak}{which variable holds the sample peak}

\item{imf}{which variable holds the IMF information}

\item{e_value}{the e-values}

\item{emf}{which variable holds the EMF information}

\item{sample}{which variable holds the sample}

\item{observed_mz}{which variable holds the observed M/Z}

\item{assigned_mz}{which variable holds the assigned M/Z}

\item{height}{which variable holds the height / intensity}

\item{other_cols}{which other columns should be kept?}

\item{progress}{should the progress be shown?}

\item{remove_secondary}{remove the seconardy assignments? (default is TRUE)}
}
\value{
list of matrices and a data.frame
}
\description{
Extract the data from a complete set of assignments.
}
