% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confident_sd.R
\name{extract_coefficient_data}
\alias{extract_coefficient_data}
\title{get coefficient data}
\usage{
extract_coefficient_data(zip_file_list)
}
\arguments{
\item{zip_file_list}{the list of zip files}
}
\value{
list of lists
}
\description{
Given a set of zip files, find the associated json metadata files, and extract the
coefficient information from them.
}
