% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfs.R
\name{extract_emf_location}
\alias{extract_emf_location}
\title{Extract Single Location}
\usage{
extract_emf_location(emf_list_location)
}
\arguments{
\item{emf_list_location}{the list of EMF locations}
}
\description{
Given a list of EMF locations, we would like to get a single location for each EMF,
as by definition an EMF often has at least two IMFs.
}
