% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfs.R
\name{extract_emf_mz}
\alias{extract_emf_mz}
\title{Extract Single M/Z}
\usage{
extract_emf_mz(emf_list_mz)
}
\arguments{
\item{emf_list_mz}{the list of EMF M/Zs}
}
\description{
Given a list of EMF M/Z, we would like to get a single M/Z for each EMF,
as by definition an EMF often has at least two IMFs.
}
