% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{extract_emf_propensity}
\alias{extract_emf_propensity}
\title{extract EMF propensity}
\usage{
extract_emf_propensity(emf_intensity)
}
\arguments{
\item{emf_intensity}{list of EMF intensities}
}
\value{
matrix
}
\description{
If one wants to be able to filter EMF based data for those EMFs that show up in a particular number
of samples, one needs to be able to count which samples the EMF appears in. This function extracts
creates a matrix where ones note the samples that contained the EMF.
}
