% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{extract_first_EMF_height}
\alias{extract_first_EMF_height}
\title{extract EMF first rows}
\usage{
extract_first_EMF_height(emf_heights)
}
\arguments{
\item{emf_heights}{list of EMF heights}
}
\value{
matrix
}
\description{
If one wants to be able to filter EMF based data for those EMFs that show up in a particular number
of samples, one needs to be able to count which samples the EMF appears in. This function extracts
the first row of \code{height}s from each EMF into a single matrix.
}
