% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{extract_imf_emf_data}
\alias{extract_imf_emf_data}
\title{extract IMF or EMF level data}
\usage{
extract_imf_emf_data(emfs, by = "EMF")
}
\arguments{
\item{emfs}{list of EMFs}

\item{by}{\code{EMF} (default) or \code{IMF}}
}
\value{
list with \code{height}, \code{mz}, and \code{info}
}
\description{
Extract either IMF or EMF level data from the extracted EMFs. Which is controlled by the \code{by} argument.
}
\details{
For \code{EMF}, will be a list corresponding to each EMF from the sudo EMFs, for \code{IMF}, will be
matrices and a data.frame.
}
