% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_assignments.R
\name{extract_peaks}
\alias{extract_peaks}
\title{extract peaks}
\usage{
extract_peaks(assignment_list, .pb = NULL)
}
\arguments{
\item{assignment_list}{list of assignments, one for each sample, generated by \code{read_smirfe_assignment}}

\item{.pb}{a progress bar object from either dplyr or knitrProgressBar}
}
\value{
list with height, emf, peak, and type, and emf_imf list
}
\description{
To be computably useful, the peak information for each IMF needs to be in a
matrix type format. This takes a list of peak-IMF data.frames, and generates
a list of matrices to hold the data, where the matrix has IMF rows and sample
columns.
}
