% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confident_sd.R
\name{find_confident_frequency_sd}
\alias{find_confident_frequency_sd}
\title{find confident frequency sd}
\usage{
find_confident_frequency_sd(
  assigned_data,
  low_evalue_cutoff = 0.1,
  low_mz_cutoff = 600,
  remove_elements = "S"
)
}
\arguments{
\item{assigned_data}{list of assignment data}

\item{low_evalue_cutoff}{upper bound for what is considered a \emph{confident} assignment}

\item{low_mz_cutoff}{the upper bound for \emph{confident} assignments in M/Z}
}
\value{
data.frame with sd values
}
\description{
Given a list of EMF assignment data from SMIRFE, attempts to find \emph{confident}
EMF assignments, and use those combined with scan level frequency information
to determine a suitable cutoff across sample assignments.
}
