% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_assignments.R
\name{find_duplicate_peaks}
\alias{find_duplicate_peaks}
\title{find duplicate peaks}
\usage{
find_duplicate_peaks(extracted_data, .pb = NULL)
}
\arguments{
\item{extracted_data}{generated by \code{extract_peaks}}

\item{.pb}{a progress object, should be equal to the number of samples used.}
}
\value{
deduplicated list
}
\description{
find duplicate peaks
}
