% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_to_mz_cutoff.R
\name{fit_predict_mz_cutoff}
\alias{fit_predict_mz_cutoff}
\title{fit / predict M/Z}
\usage{
fit_predict_mz_cutoff(mz_sd, multiplier = 2, step_size = 0.5)
}
\arguments{
\item{mz_sd}{data.frame of Value and Index}

\item{multiplier}{what multiplier of the base fit should be used? (default = 2)}

\item{step_size}{the increment to use going from low to high values of M/Z}
}
\value{
data.frame of Index and Value
}
\description{
Given a M/Z SD's generated by \code{get_mz_after_freq}, fit a model, create smaller data,
and predict values for M/Z based EMF voting.
}
