% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_assignments.R
\name{get_assigned_peak_info}
\alias{get_assigned_peak_info}
\title{peak mz, height, assignments}
\usage{
get_assigned_peak_info(peak_list, variables = c("ObservedMZ", "Height"),
  summaries = "Mean", include_non_primary = TRUE,
  keep_peaks = "assignments")
}
\arguments{
\item{peak_list}{the peak list from reading some JSON}

\item{variables}{which variables to extract (default: ObservedMZ, Height)}

\item{summaries}{which summary variables to extract (default: Mean)}

\item{include_non_primary}{include non-primary assignments (default: TRUE)}

\item{keep_peaks}{keep "both" original and assigned peaks, "assignments" only, or original "peaks"}
}
\value{
data.frame
}
\description{
Extracts desired peak variables and assignment information from a peak list,
the peak list is typically read in from a JSON output file of assignments
generated by SMIRFE.
}
