% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfs.R
\name{get_c13_incorporation}
\alias{get_c13_incorporation}
\title{Count Carbon 13s}
\usage{
get_c13_incorporation(imf)
}
\arguments{
\item{imf}{a character vector of isotopically resolved molecular formula's (IMFs)}
}
\value{
numeric vector
}
\description{
Counts the number of carbons in the SMIRFE IMF formula's
}
