% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_counts.R
\name{get_emf_elements}
\alias{get_emf_elements}
\title{get EMF elements}
\usage{
get_emf_elements(emf_string)
}
\arguments{
\item{emf_string}{the EMF string to extract from.}
}
\value{
data.frame with element and count
}
\description{
For an elemental molecular formula, extract the elements and their counts
}
