% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfs.R
\name{get_emf_intensity_sums}
\alias{get_emf_intensity_sums}
\title{Get EMF sums}
\usage{
get_emf_intensity_sums(split_emfs, feature_intensities, feature_location,
  multiply_C13 = FALSE)
}
\arguments{
\item{split_emfs}{list of EMF data.frames}

\item{feature_intensities}{matrix of feature (rows) by sample (column) intensities}

\item{feature_location}{matrix of M/Z values}

\item{multiply_C13}{should the intensities be multiplied by the C13 incorporation? (FALSE)}
}
\value{
list
}
\description{
Get EMF sums
}
