% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_counts.R
\name{get_imf_elements}
\alias{get_imf_elements}
\title{get IMF element counts}
\usage{
get_imf_elements(imf_string)
}
\arguments{
\item{imf_string}{the imf string to use}
}
\value{
data.frame with element and count.
}
\description{
Given an isotopic molecular formula (IMF), extract the counts for each element.
}
