% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_to_mz_cutoff.R
\name{get_mz_after_freq}
\alias{get_mz_after_freq}
\title{get M/Z after Frequency}
\usage{
get_mz_after_freq(assigned_set, freq_cutoff, set_id = NULL)
}
\arguments{
\item{assigned_set}{a set of samples defined by their common frequency conversion}

\item{freq_cutoff}{the cutoff used to exclude peaks in frequency space}

\item{set_id}{an optional identifier to identify which set of IMFs we have}
}
\value{
data.frame of Index (M/Z) and Value (SD) and \emph{set} (if set_id is supplied)
}
\description{
The Frequency based cutoffs should be the most reliable \strong{within} a set of
samples. However, they do not translate well across sample sets. Therefore,
a method is to split up the samples, run EMF voting using the derived frequency cutoff,
and then fit a generalized additive model to their M/Z standard deviations for
each IMF.
}
