% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfs.R
\name{group_emfs_by_peaks}
\alias{group_emfs_by_peaks}
\title{EMFs to Peaks}
\usage{
group_emfs_by_peaks(peak_info, peak_var = "peak",
  emf_var = "complete_EMF", grouping_var = "grouped_EMF")
}
\arguments{
\item{peak_info}{data.frame mapping peaks to EMFs}

\item{peak_var}{which column has the peak information}

\item{emf_var}{which column has the EMF information}

\item{grouping_var}{what should the grouping variable be called}
}
\value{
data.frame
}
\description{
If we want to work with EMFs, we should first determine what are \code{unique} EMFs
based on the mapping between peaks and EMFs. This function adds a new column
to the data.frame, \code{grouped_EMF}
}
