% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_to_mz_cutoff.R
\name{grouped_mz_after_freq}
\alias{grouped_mz_after_freq}
\title{mz after frequency by group}
\usage{
grouped_mz_after_freq(grouped_samples, freq_cutoff)
}
\arguments{
\item{grouped_samples}{the split sample assigned_data}

\item{freq_cutoff}{the frequency cutoff to use}
}
\value{
data.frame with Index, Value, and set
}
\description{
Given groups of samples in a list, run each one independently using frequency
cutoffs to calculate M/Z cutoffs
}
