% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming.R
\name{hamming}
\alias{hamming}
\title{hamming distance}
\usage{
hamming(X, Y)
}
\arguments{
\item{X}{the first matrix}

\item{Y}{the second matrix}
}
\value{
double
}
\description{
calculates the hamming distance between two matrices or between columns
of a matrix using matrix multiplication. Copied from \href{https://johanndejong.wordpress.com/2015/10/02/faster-hamming-distance-in-r-2/}{Johan DeJong's blog},
copyright is ascribed to them.
}
