% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfs.R
\name{keep_emfs_with_features}
\alias{keep_emfs_with_features}
\title{keep EMFs with matching features}
\usage{
keep_emfs_with_features(features, emf_info)
}
\arguments{
\item{features}{a set of interesting features}

\item{emf_info}{data.frame of EMFs, including the lipid class information}
}
\value{
list
}
\description{
Given a set of features of interest, and a set of decided EMFs, goes through
and returns the full EMFs that contain any of the features of interest. Also
attempts to return as few classes of EMFs as possible.
}
