% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keeping_adducts.R
\name{keep_imf_adducts}
\alias{keep_imf_adducts}
\title{extracts IMF adducts}
\usage{
keep_imf_adducts(assignments, imf = "complete_IMF",
  adduct = "adduct_IMF")
}
\arguments{
\item{assignments}{a data.frame of assignments}

\item{imf}{which is the IMF column we want to keep}

\item{adduct}{which variable is holding the adduct information?}
}
\value{
data.frame
}
\description{
Although technically all complete_EMFs are equivalent, they might not be
due to differences in the isotopologue adduct, for example Na vs K, 41K vs
39K, H vs K vs Na. Therefore, we need to be able to save which adduct
belongs with each IMF.
}
