% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{match_imf_by_frequency}
\alias{match_imf_by_frequency}
\title{match peaks to IMFs by frequency}
\usage{
match_imf_by_frequency(imf_2_peak, unknown_peaks, peak_location,
  difference_cutoff)
}
\arguments{
\item{imf_2_peak}{data.frame with IMF information (see details)}

\item{unknown_peaks}{data.frame of peak information}

\item{peak_location}{data.frame with \code{Sample_Peak} and \code{Value}, where \code{Value} is \code{ObservedFrequency}}

\item{difference_cutoff}{the maximum difference an unknown peak can have to any of the known peaks}
}
\value{
data.frame
}
\description{
Given a data.frame of IMFs to peaks for a given EMF, attempts to match unknown peaks
to IMFs based on differences in frequency to the unknown peaks.
}
