% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{match_imf_by_mz}
\alias{match_imf_by_mz}
\title{match peaks to IMFs by M/Z}
\usage{
match_imf_by_mz(imf_2_peak, unknown_peaks, peak_mz)
}
\arguments{
\item{imf_2_peak}{data.frame with IMF information (see details)}

\item{unknown_peaks}{data.frame of peak information}

\item{peak_mz}{data.frame with \code{Sample_Peak} and \code{Value}, where \code{Value} is \code{ObservedMZ}}
}
\value{
data.frame
}
\description{
Given a data.frame of IMFs to peaks for a given EMF, attempts to match unknown peaks
to IMFs based on differences in average M/Z to the unknown peaks.
}
