% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_emf_voting.R
\name{old_add_location_intensity}
\alias{old_add_location_intensity}
\title{add peak location and intensity}
\usage{
old_add_location_intensity(
  emfs,
  peak_data,
  location = ObservedMZ,
  intensity = Height
)
}
\arguments{
\item{emfs}{list of EMFs}

\item{peak_data}{data.frame of peak data}

\item{location}{which variable to use for location (unquoted)}

\item{intensity}{which variable to use for intensity (unquoted)}
}
\value{
list
}
\description{
Given a set of EMF peak matrices and a data.frame of various locations and
intensities, associates a matrix of each of location and intensity directly
to the EMF peak matrix. Can be used to extract alternatives to the default
of Height and ObservedMZ.
}
