% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_emf_voting.R
\name{old_extract_imf_emf_data}
\alias{old_extract_imf_emf_data}
\title{extract IMF or EMF level data}
\usage{
old_extract_imf_emf_data(emfs, by = "EMF")
}
\arguments{
\item{emfs}{list of EMFs}

\item{by}{\code{EMF} (default) or \code{IMF}}
}
\value{
list with \code{intensity}, \code{location}, and \code{info}
}
\description{
This function is deprecated, and kept around merely for use in some old code.
Extract either IMF or EMF level data from the extracted EMFs. Which is controlled by the \code{by} argument.
}
\details{
For \code{EMF}, will be a list corresponding to each EMF from the sudo EMFs, for \code{IMF}, will be
matrices and a data.frame.
}
