% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_assignments.R
\name{one_peak_from_imfs}
\alias{one_peak_from_imfs}
\title{single peaks}
\usage{
one_peak_from_imfs(peak_assignments)
}
\arguments{
\item{peak_assignments}{data.frame of peak assignments}
}
\value{
list of IMFs, each with a single peak
}
\description{
given a data.frame of peak assignments, go through and choose a single peak
where needed. Trying to avoid running \code{choose_peak} on every single IMF.
}
