% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{remove_duplicates_across_semfs}
\alias{remove_duplicates_across_semfs}
\title{remove duplicate peaks across sudo EMFs}
\usage{
remove_duplicates_across_semfs(
  chosen_emfs,
  all_gemfs,
  scan_level_location,
  peak_location,
  difference_cutoff,
  keep_ratio = 0.9,
  min_gemfs = 1
)
}
\arguments{
\item{chosen_emfs}{merged EMFs}

\item{all_gemfs}{the grouped_EMFs}

\item{peak_location}{the data.frame of peak_location}

\item{difference_cutoff}{tolerance for matched peaks for \code{choose_emf}}

\item{keep_ratio}{for \code{choose_emf}}

\item{min_gemfs}{minimum number of grouped EMFs for each sudo EMF}
}
\value{
list of sudo EMFs after voting
}
\description{
Given a set of sudo EMFs, detects duplicate peaks and removes them, and carries
out EMF voting on what is left. Should only be used after an initial round of
merging.
}
