% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_samples.R
\name{remove_samples}
\alias{remove_samples}
\title{remove samples}
\usage{
remove_samples(sample_df, logical_condition = NULL, reason = NULL)
}
\arguments{
\item{sample_df}{the sample data.frame to modify}

\item{logical_condition}{which samples are to be removed}

\item{reason}{why should the samples be removed}
}
\value{
data.frame
}
\description{
Provides ways for the user to note in a sample data.frame which samples were
kept or removed, and for which reasons.
}
