% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_assignments.R
\name{score_filter_assignments}
\alias{score_filter_assignments}
\title{score and filter assignments}
\usage{
score_filter_assignments(
  assignments,
  score_calculation = 1 - e_value,
  score_weight = 1,
  filter_conditions = ObservedMZ <= 1600,
  high_correlation = 0.5,
  emf_weight = NULL,
  remove_highsd = TRUE,
  remove_contaminants = TRUE
)
}
\arguments{
\item{assignments}{a character or list object}

\item{score_calculation}{how to calculate the score}

\item{score_weight}{multiplier to apply to the scores (default = 1)}

\item{filter_conditions}{how the data should be filtered}

\item{high_correlation}{what is considered high correlation with scan}

\item{emf_weight}{an optional data.frame of emf_weights}

\item{remove_highsd}{should the high standard deviation peaks be removed}

\item{remove_contaminants}{should peaks that were marked as contaminants be removed}
}
\value{
data.frame of assignments with scores added
}
\description{
Given assignments (either a filename or list), calculate scores, and filter
the assignments by M/Z
}
\seealso{
weight_lipid_classifications
}
