% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_internal_map}
\alias{set_internal_map}
\title{pick map_at enumerator}
\usage{
set_internal_map(map_function = NULL)
}
\arguments{
\item{map_function}{which function to use, assigns it to an internal object}
}
\description{
Allows the user to set which enumerator is being used internally in the functions.
}
