% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting_emfs.R
\name{weight_lipid_classifications}
\alias{weight_lipid_classifications}
\title{create lipids of interest}
\usage{
weight_lipid_classifications(lipid_classification, lipid_weight = 1,
  not_lipid_weight = 1, other_weight = 1, all_emfs = NULL)
}
\arguments{
\item{lipid_classification}{data.frame of lipid classifications}

\item{lipid_weight}{what weighting to use for \strong{lipids}}

\item{not_lipid_weight}{what weighting to use for \strong{not_lipids} (metabolites)}

\item{other_weight}{what value to use for other things}

\item{all_emfs}{the full list of EMFs that we tried to classify}
}
\value{
data.frame
}
\description{
Given a data.frame of lipid classifications, creates a data.frame which defines
the \code{isotopologue_EMF} as either of interest or not (1 / 0), which can
subsequently be used to define if the score for that EMF should be modified or not.
}
\seealso{
import_emf_classifications
}
