% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_assignments.R
\name{choose_single_peak}
\alias{choose_single_peak}
\title{choose a peak}
\usage{
choose_single_peak(
  mz_diff_data,
  imf = "IMF",
  sample_peak = "Sample_Peak",
  sample = "Sample",
  data_col = "Assignment_Data"
)
}
\arguments{
\item{mz_diff_data}{the data.frame with mz_errors}

\item{imf}{which variable are the IMFs listed}

\item{sample_peak}{which variable contains the sample peak information}

\item{sample}{which variable has the sample information}

\item{data_col}{which variable has the actual mz mass error}
}
\value{
data.frame Peaks and Samples
}
\description{
When there are multiple peaks in a sample, we need to choose one based on
the chosen representative IMFs. See \strong{Details} for more information on how
that decision is made.
}
\details{
If there are multiple peaks for each sample, picks the one with
the smallest M/Z error with the provided isotopologues.
}
