% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{extract_imf_emf_data}
\alias{extract_imf_emf_data}
\title{extract IMF or EMF level data}
\usage{
extract_imf_emf_data(
  extracted_emfs,
  intensity = Height,
  location = ObservedMZ,
  by = "EMF",
  scanlevel = FALSE
)
}
\arguments{
\item{extracted_emfs}{list of data from \code{extract_assigned_data}}

\item{intensity}{which intensity measure should be used?}

\item{location}{which location should be used?}

\item{by}{\code{EMF} (default) or \code{IMF}}

\item{scanlevel}{include scan level data as well? (default is FALSE)}
}
\value{
list with \code{intensity}, \code{location}, and \code{info}
}
\description{
Extract either IMF or EMF level data from the extracted EMFs. Which is controlled by the \code{by} argument.
}
\details{
For \code{EMF}, will be a list corresponding to each EMF from the sudo EMFs, for \code{IMF}, will be
matrices and a data.frame.
}
