% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{merge_duplicate_semfs}
\alias{merge_duplicate_semfs}
\title{merge sudo EMFs with duplicate peaks}
\usage{
merge_duplicate_semfs(chosen_emfs, all_gemfs, peak_frequency,
  frequency_match_cutoff, keep_ratio = 0.9)
}
\arguments{
\item{chosen_emfs}{sudo EMFs from \code{choose_emf}}

\item{all_gemfs}{the grouped_EMFs}

\item{peak_frequency}{frequency information for each and every peak}

\item{frequency_match_cutoff}{how close do peaks need to be to each other}

\item{keep_ratio}{the ratio used to keep other highly voted EMFs}
}
\value{
list
}
\description{
After choosing EMFs, there may be sudo EMFs that share peaks. This is not kosher, there
has to be a 1-1 mapping of peaks to sudo EMFs. This function merges those that share

50% of peaks from either sudo EMF, and carries out voting again.
After merging, it checks for duplicates again, and removes those peaks that still
span multiple sudo EMFs, and then votes on them again.
}
