% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emf_voting.R
\name{remove_duplicates_across_semfs}
\alias{remove_duplicates_across_semfs}
\title{remove duplicate peaks across sudo EMFs}
\usage{
remove_duplicates_across_semfs(chosen_emfs, all_gemfs, peak_frequency,
  frequency_match_cutoff, keep_ratio = 0.9)
}
\arguments{
\item{chosen_emfs}{merged EMFs}

\item{all_gemfs}{the grouped_EMFs}

\item{peak_frequency}{the data.frame of peak_frequency}

\item{frequency_match_cutoff}{tolerance for matched peaks for \code{choose_emf}}

\item{keep_ratio}{for \code{choose_emf}}
}
\value{
list of sudo EMFs after voting
}
\description{
Given a set of sudo EMFs, detects duplicate peaks and removes them, and carries
out EMF voting on what is left. Should only be used after an initial round of
merging.
}
